📖(service workers): add offline capabilities using service workers
Awesome! This gets the docs caching and using service workers.
The caching right now is a bit full-on and it caches the entire site on load so will need some tweaking, but it's pretty cool!

Both HTML and other requests are always cached.
HTML always waits for and shows fresh data, only falling back if the page can't be fetched.
CSS etc tries cache first every time, but also grabs and caches a fresh copy for the next load.